Allow configuring infoview width
This commit alows you to configure the width of the info window as well
as the height. It also sets the filetype of the info window (as
lean.nvim does), which allows for future syntax highlighting and easier
scriptability using autocommands.

I also cleaned up a few things that HLS was complaining about, but I'm
happy to save those changes for another commit.